# hw12_13_14_15_16_calendar/api/openapi.yaml
openapi: 3.0.3
info:
  title: Calendar API
  description: HTTP API для сервиса календаря
  version: 1.0.0
servers:
  - url: http://localhost:8888
    description: Local development server

components:
  schemas:
    Event:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        user_id:
          type: integer
          format: int64
        description:
          type: string
          nullable: true
        notify_before:
          type: integer
          format: int32
          nullable: true
          description: За сколько минут до события отправлять уведомление
      example:
        id: 1
        title: "Встреча с командой"
        start_time: "2026-03-27T10:00:00Z"
        end_time: "2026-03-27T11:00:00Z"
        user_id: 12345
        description: "Обсуждение спринта"
        notify_before: 15
    
    CreateEventRequest:
      type: object
      required:
        - title
        - start_time
        - end_time
        - user_id
      properties:
        title:
          type: string
          example: "Встреча с командой"
        start_time:
          type: string
          format: date-time
          example: "2026-03-27T10:00:00Z"
        end_time:
          type: string
          format: date-time
          example: "2026-03-27T11:00:00Z"
        user_id:
          type: integer
          format: int64
          example: 12345
        description:
          type: string
          nullable: true
          example: "Обсуждение спринта"
        notify_before:
          type: integer
          format: int32
          nullable: true
          example: 15
      example:
        title: "Встреча с командой"
        start_time: "2026-03-27T10:00:00Z"
        end_time: "2026-03-27T11:00:00Z"
        user_id: 12345
        description: "Обсуждение спринта"
        notify_before: 15
    
    UpdateEventRequest:
      type: object
      properties:
        title:
          type: string
          example: "Встреча с командой"
        start_time:
          type: string
          format: date-time
          example: "2026-03-27T10:00:00Z"
        end_time:
          type: string
          format: date-time
          example: "2026-03-27T11:00:00Z"
        user_id:
          type: integer
          format: int64
          example: 12345
        description:
          type: string
          nullable: true
          example: "Обсуждение спринта"
        notify_before:
          type: integer
          format: int32
          nullable: true
          example: 15
      example:
        title: "Встреча с командой"
        start_time: "2026-03-27T10:00:00Z"
        end_time: "2026-03-27T11:00:00Z"
        user_id: 12345
        description: "Обсуждение спринта"
        notify_before: 15
    
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Событие не найдено"
      example:
        error: "Событие не найдено"

    DeleteResponse:
      type: object
      properties:
        message:
          type: string
          example: "Событие удалено"
      example:
        message: "Событие удалено"

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT токен для аутентификации. Добавьте заголовок: Authorization: Bearer <token>"

paths:
  /api/v1/events:
    get:
      summary: Получить все события
      operationId: getEvents
      tags:
        - events
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    post:
      summary: Создать новое событие
      operationId: createEvent
      tags:
        - events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
      responses:
        '201':
          description: Событие создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Дата занята
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  
  /api/v1/events/{id}:
    get:
      summary: Получить событие по ID
      operationId: getEvent
      tags:
        - events
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Событие не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    put:
      summary: Обновить событие
      operationId: updateEvent
      tags:
        - events
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEventRequest'
      responses:
        '200':
          description: Событие обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Событие не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Дата занята
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    delete:
      summary: Удалить событие
      operationId: deleteEvent
      tags:
        - events
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Событие удалено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '404':
          description: Событие не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/events/day/{date}:
    get:
      summary: Получить события на день
      operationId: getEventsByDay
      tags:
        - events
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
          description: Дата в формате YYYY-MM-DD
          example: "2026-03-27"
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '400':
          description: Некорректный формат даты
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/events/week/{date}:
    get:
      summary: Получить события на неделю
      operationId: getEventsByWeek
      tags:
        - events
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
          description: Дата начала недели в формате YYYY-MM-DD
          example: "2026-03-27"
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '400':
          description: Некорректный формат даты
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/events/month/{date}:
    get:
      summary: Получить события на месяц
      operationId: getEventsByMonth
      tags:
        - events
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
          description: Дата начала месяца в формате YYYY-MM-DD
          example: "2026-03-27"
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '400':
          description: Некорректный формат даты
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'